Cast close code to an int
The Qt company is correct that the web socket spec states that
a close code should be between 1000 and 1015, but it's not correct
to limit this with an enum because non-standard responses are
still possible, and apparently Twitch is doing this very thing by
returning codes in the 4000s (which the spec permits but doesn't
define).

https://doc.qt.io/qt-6/qwebsocketprotocol.html#CloseCode-enum
https://www.rfc-editor.org/rfc/rfc6455.html#section-7.4.1
https://dev.twitch.tv/docs/eventsub/handling-websocket-events/#close-message

+++ eventsub.cpp
@@ -83,7 +83,7 @@ void EventSub::SocketClosed()
 {
 	static const char *TWITCH_API_OPERATION_SOCKET_CLOSED="socket closed";
 
-	switch (socket.closeCode())
+	switch (static_cast<int>(socket.closeCode()))
 	{
 	case QWebSocketProtocol::CloseCodeNormal:
 		break;
